docs: clarify develop vs main branch
Since the default branch is `develop`, when users clone the repo, it
will be in this branch by default. This change clarifies in the README
that if they want the latest stable code, they should switch to the
`main` branch.

Add information about updated the deno submodule when pulling into a
previously checked out repo.

Resolves: #231