feat: add check-checker options to Clarinet.toml
The options in the check-checker are off by default. By explicitly
adding the options in the Clarinet.toml that gets generated, it makes it
easy for a developer to turn those options on. This also provides a
place, that developers should notice, to include some comments about
what these options do.