fix: crash on `clarinet new contract`
After the addition of `repl` in `ProjectManifest` in a5695025, calling
`toml::to_string` on this struct caused a `ValueAfterTable` error. This
problem is handled in the display of the value, so this alternative
method of generating the string avoids the problem.

This is a quick fix, but this code will soon be changed to solve #246.