feat: add a simple easter egg to the REPL
When a developer types `/-/` in the REPL (`clarinet console`), it will
pull a message from a GitHub gist and display it. This is just a super
quick PoC to decide if we want to include it. The downside is that it
requires pulling `reqwest` into `clarity-repl`.

An alternative is to only change the message once per release, then it
doesn't need to pull anything from the Internet, but it could limit the
fun that the marketing team can have with it.