feat: Add `epoch` field to batch in deployment plans
If an `epoch` is specified for a batch of transactions in the deployment
plan, then this batch will not be deployed until that epoch is live.
Note that the batches are still processed in order, so if a batch that
specifies `2.1` is followed by a batch that specifies `2.05`, that
second batch will run in 2.1.

Example:

```yaml
---
id: 0
name: Devnet deployment
network: devnet
stacks-node: "http://localhost:20443"
bitcoin-node: "http://devnet:devnet@localhost:18443"
plan:
  batches:
    - id: 0
      epoch: 2.05
      transactions:
        - contract-publish:
            contract-name: early
            expected-sender: ST1PQHQKV0RJXZFY1DGX8MNSNYVE3VGZJSRTPGZGM
            cost: 830
            path: contracts/early.clar
            anchor-block-only: true
    - id: 1
      epoch: 2.1
      transactions:
        - contract-publish:
            contract-name: test
            expected-sender: ST1PQHQKV0RJXZFY1DGX8MNSNYVE3VGZJSRTPGZGM
            cost: 1180
            path: contracts/test.clar
            anchor-block-only: true
            clarity-version: 2
```