chore: update default epoch for contracts
This is the epoch used when a contract in a project's Clarinet.toml does
not have an epoch specified. Previously this defaulted to 2.0, which can
create a scenario where a requirement is deployed in 2.05, then the
contract that depends on it is deployed in 2.0, causing an error because
the ordering is off. This is now set to default to 2.05, but should
always be updated to whatever is the current epoch.