fix: use correct clarity version for requirements
We can retrieve the clarity version while we retrieve the contract,
since it is now returned from the API. If the version is not specified,
then we can compute the default clarity version for the current epoch
using the block height of the deploy transaction.

Also fill in the exact block number for testnet 2.1 activation.